refactor delegators count
