+++ src/types/kv.ts
@@ -1,0 +1,5 @@
+export interface ActiveValidatorsKV {
+  totalDelegatorsCount?: string;
+  votingPower?: string;
+  updatedAt?: string;
+}
